VMM/PDMAsyncCompletion: Use 64-bit types for the bandwidth group limits, ticketref:22328 [missing file]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108461 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/ConsoleImpl.cpp
@@ -6616,7 +6616,7 @@ HRESULT Console::i_onBandwidthGroupChange(IBandwidthGroup *aBandwidthGroup)
                         int vrc = VINF_SUCCESS;
                         if (enmType == BandwidthGroupType_Disk)
                             vrc = ptrVM.vtable()->pfnPDMR3AsyncCompletionBwMgrSetMaxForFile(ptrVM.rawUVM(), strName.c_str(),
-                                                                                            (uint32_t)cMax);
+                                                                                            (uint64_t)cMax);
 #ifdef VBOX_WITH_NETSHAPER
                         else if (enmType == BandwidthGroupType_Network)
                             vrc = ptrVM.vtable()->pfnPDMR3NsBwGroupSetLimit(ptrVM.rawUVM(), strName.c_str(), cMax);
